Struct glib_signal :: SignalFlags

pub struct SignalFlags { /* private fields */ }

Implementations

impl SignalFlags

pub const RUN_FIRST : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_RUN_FIRST as u32,}

pub const RUN_LAST : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_RUN_LAST as u32,}

pub const RUN_CLEANUP : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_RUN_CLEANUP as u32,}

pub const NO_RECURSE : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_NO_RECURSE as u32,}

pub const DETAILED : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_DETAILED as u32,}

pub const ACTION : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_ACTION as u32,}

pub const NO_HOOKS : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_NO_HOOKS as u32,}

pub const MUST_COLLECT : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_MUST_COLLECT as u32,}

pub const DEPRECATED : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_DEPRECATED as u32,}

pub const ACCUMULATOR_FIRST_RUN : SignalFlags = Self{ bits: gobject_ffi::G_SIGNAL_ACCUMULATOR_FIRST_RUN as u32,}

pub const fn empty () -> SignalFlags

Returns an empty set of flags.

pub const fn all () -> SignalFlags

Returns the set containing all flags.

pub const fn bits (&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits (bits: u32 ) -> Option < SignalFlags >

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate (bits: u32 ) -> SignalFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked (bits: u32 ) -> SignalFlags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty (&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all (&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects (&self, other: SignalFlags ) -> bool

Returns true if there are flags common to both self and other .

pub const fn contains (&self, other: SignalFlags ) -> bool

Returns true if all of the flags in other are contained within self .

pub fn insert (&mut self, other: SignalFlags )

Inserts the specified flags in-place.

pub fn remove (&mut self, other: SignalFlags )

Removes the specified flags in-place.

pub fn toggle (&mut self, other: SignalFlags )

Toggles the specified flags in-place.

pub fn set (&mut self, other: SignalFlags , value: bool )

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection (self, other: SignalFlags ) -> SignalFlags

Returns the intersection between the flags in self and other .

Specifically, the returned set contains only the flags which are present in both self and other .

This is equivalent to using the & operator (e.g. ops::BitAnd ), as in flags & other .

pub const fn union (self, other: SignalFlags ) -> SignalFlags

Returns the union of between the flags in self and other .

Specifically, the returned set contains all flags which are present in either self or other , including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr ), as in flags | other .

pub const fn difference (self, other: SignalFlags ) -> SignalFlags

Returns the difference between the flags in self and other .

Specifically, the returned set contains all flags present in self , except for the ones present in other .

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub ), as in flags - other .

pub const fn symmetric_difference (self, other: SignalFlags ) -> SignalFlags

Returns the symmetric difference between the flags in self and other .

Specifically, the returned set contains the flags present which are present in self or other , but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other .

This is equivalent to using the ^ operator (e.g. ops::BitXor ), as in flags ^ other .

pub const fn complement (self) -> SignalFlags

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self , but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self )

This is equivalent to using the ! operator (e.g. ops::Not ), as in !flags .